Create lamp_test
Simple code for a lamp that change color after read a RFID tag (each tag represents a color). It publish this color information in a MQTT server that subscribe it in another lamp arround the world. The two lamps always get the same color, don't matter where it is arround the world.